Bridge discrete line segments separated by null values.  Has the effect of unifying disparate segments into a single curve.  Introducing bridgeNulls opt to enable bridging, disabled by default.
The theme for a bridge is based on the stroke for the line, with dashes applied.  Area plots remain hollow under bridges.

Bridges are independent line segments that visually connect actual line segments.  Tensioning will not have any effect on the bridges, nor will the bridge affect the grounded segments.

For simplicity and deference to the current render model, the bridge line will slightly overlap the marker at the start of the bridge.  In most cases, this will not be noticeable.