Adding the ability to specify a full size event mask for a Bars/Columns plot.  The mask enables the user to easily interact with small plot elements.  Introducing fullMask to enable the mask, it is disabled by default.
For example, with a chart scaled from 0 to 100, a bar with a value of 1 is going to be tinny compared to the size of the plot.  In order for a user to interact with that bar, they must mouse over the bar as it is rendered.  With fullMask enabled, the user can interact with the bar as if it were rendered with a value of 100, creating a much larger and more consistent event mask.